+++ composeApp/src/androidMain/kotlin/pages/LoginMenu.android.kt
@@ -5,6 +5,8 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.DropdownMenu
+import androidx.compose.material.DropdownMenuItem
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
@@ -39,10 +41,10 @@ actual fun LoginMenu(
 
     )
     Column {
-        AsyncImage(
-            model = logo,//turn into bitmap
-            contentDescription = "Logo"
-        )
+//        AsyncImage(
+//            model = logo,//turn into bitmap
+//            contentDescription = "Logo"
+//        )
         Text(
             text = "Login",
             fontSize = 45.sp,
@@ -92,26 +94,20 @@ actual fun LoginMenu(
             DropdownMenu(expanded = compDD, onDismissRequest = { compDD = false; },modifier= Modifier.background(color = getCurrentTheme().onSurface)) {
                 DropdownMenuItem(
                     onClick = { comp.value = "Bonney Lake"; compDD = false; compKey = tbaMatches[0]},
-                    text = { Text(text = "Bonney Lake", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
-                )
+                ){ Text(text = "Bonney Lake", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
                 DropdownMenuItem(
                     onClick = { comp.value = "Lake Sammamish"; compDD = false; compKey = tbaMatches[1]},
-                    text = { Text(text = "Lake Sammamish", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
-                )
+                ){ Text(text = "Lake Sammamish", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
                 DropdownMenuItem(
                     onClick = { comp.value = "Salem"; compDD = false; compKey = tbaMatches[2]},
-                    text ={ Text(text = "Salem", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
-                )
+                ){ Text(text = "Salem", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
                 DropdownMenuItem(
                     onClick = { comp.value = "Houston"; compDD = false; compKey = tbaMatches[3]},
-                    text ={ Text(text = "Houston", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
-                )
+                ){ Text(text = "Houston", color = getCurrentTheme().onPrimary,modifier= Modifier.background(color = getCurrentTheme().onSurface)) }
             }
 
         }
-        Divider(
-            color = defaultPrimaryVariant,
-        )
+        HorizontalDivider(color = defaultPrimaryVariant)
         OutlinedButton(
             onClick = {
                 if (comp.value != "" && scoutName.value != "")
