+++ composeApp/src/androidMain/kotlin/pages/PitsScoutMenu.android.kt
@@ -33,6 +33,7 @@ import com.bumble.appyx.navigation.modality.BuildContext
 import com.bumble.appyx.navigation.node.Node
 import composables.CheckBox
 import composables.Profile
+import composables.download
 import defaultOnPrimary
 import defaultPrimaryVariant
 import org.example.project.ComposeFileProvider
@@ -392,7 +393,7 @@ actual class PitsScoutMenu actual constructor(
             Row{
                 OutlinedButton(onClick = { if (photoArray.size >= 1) { robotCard = true }}) { Text(text = "Submit", color = defaultOnPrimary) }
                 OutlinedButton(onClick = { robotCard = false }) { Text(text = "Close", color = defaultOnPrimary) }
-                OutlinedButton(onClick = {}) { Text(text = "Download", color = defaultOnPrimary) }
+                OutlinedButton(onClick = {download(context,photoArray,scoutedTeamNumber,photoAmount)}) { Text(text = "Download", color = defaultOnPrimary) }
                 OutlinedButton(onClick = { backStack.push(RootNode.NavTarget.MainMenu) }) { Text(text = "Back", color = defaultOnPrimary)
                 }
             }
