+++ gradle/libs.versions.toml
@@ -28,16 +28,12 @@ okhttp = "4.12.0"
 slf4jNop = "2.1.0-alpha1"
 uiDesktop = "1.5.11"
 material3 = "1.2.0"
-#material3 = "1.2.0"
-material3Desktop = "1.2.0"
 webcamCapture = "0.3.10"
-material3Android = "1.2.0"
 core = "1.41.0"
 filamentAndroid = "1.17.1"
 cardview = "1.0.0"
 volley = "1.2.1"
 databindingCompiler = "8.2.2"
-#material3Android = "1.2.0"
 uiUnitAndroid = "1.6.1"
 
 [libraries]
@@ -72,9 +68,8 @@ qrcode-kotlin = { module = "io.github.g0dkar:qrcode-kotlin", version.ref = "qrco
 slf4j-nop = { module = "org.slf4j:slf4j-nop", version.ref = "slf4jNop" }
 ui-desktop = { module = "org.jetbrains.compose.ui:ui-desktop", version.ref = "uiDesktop" }
 androidx-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "material3" }
-androidx-material3-desktop = { group = "androidx.compose.material3", name = "material3-desktop", version.ref = "material3Desktop" }
+androidx-material3-desktop = { group = "androidx.compose.material3", name = "material3-desktop", version.ref = "material3" }
 androidx-databinding-compiler = { group = "androidx.databinding", name = "databinding-compiler", version.ref = "databindingCompiler" }
-#androidx-material3-android = { group = "androidx.compose.material3", name = "material3-android", version.ref = "material3Android" }
 androidx-ui-unit-android = { group = "androidx.compose.ui", name = "ui-unit-android", version.ref = "uiUnitAndroid" }
 webcam-capture = { module = "com.github.sarxos:webcam-capture", version.ref = "webcamCapture" }
 core = { group = "com.google.ar", name = "core", version.ref = "core" }
